move test to own file
