+++ tests/test_rand.py
@@ -1,0 +1,10 @@
+from ddtrace.internal import _rand
+
+
+def test_random(self):
+    m = set()
+    for i in range(0, 2**16):
+        n = _rand.rand64bits()
+        assert 0 <= n <= 2**64 - 1
+        assert n not in m
+        m.add(n)
